requirements.txt
pandas
openai
python-dotenv
langchain
langchain-openai
langchain-core
pydantic